refactor(matrices): move domain-handling code to MatrixBase
